Switch to using a separate command instead of overloading `loaddata`.
